Don't use Graphics.DrawImageUnscaled, fixes incorrect image size
The name of the method is misleading, it actually draws the image so that it has the same physical size (in inches) when displayed on the screen and is effected by the dpi in the source image and the dpi of the monitor. Just use DrawImage to keep the image file's dimensions the same.