Added new IMediaViewAction and associated interfaces, to allow actions to be added to the info menu when the cursor is on a group of media items (e.g. an Album, or some recordings grouped under the same name). For the appropriate xml, see UI\UiComponents\Media\plugin.xml under MediaViewActions.
This new facility has been used to add a "Delete these recordings" to the info menu when on a group of TV recordings, and to reimplement "Add all to playlist" when pressing info on a filter group, so it just adds that filter group, not the whole screenful of items.

Note that, for this to work, the old bindings in all skins MediaStyles.xaml to ShowMediaItemActions need to be changed to call ShowMediaItemActionsEx, and to pass the ListItem instead of its MediaItem ("{Binding MediaItem}" becomes just "{Binding}").