MP2-871: Trigger update of media list directly in message handler so the model is requested before the system is shutting down in the case where a close has been requested from the system state dialog to avoid the model being requested after it has been freed by the WorkflowManager.
When a close is requested from the dialog, first the dialog is closed, so navigation returns to the home state, then the system begins shutting down, so we can't tell that the system is shutting down at the point the navigation complete message handler is called and avoid requesting the model altogether. Instead we just make sure to request the model in the message handlers call stack, which is guaranteed to be called before the Workflowmanager starts freeing models.