MP2-672: Remove unnecessary InstallConditions, DetectCondition is the correct way to handle determining whether to install a package.
If an InstallCondition evaluates to false then Burn will attempt to uninstall the package, the InstallConditions were based on whether the package is already installed and would evaluate to true if not and false if so. This meant that Burn would set the package to be uninstalled if it was already installed, this behaviour was masked because all relevant packages are marked as permanent so are never uninstalled, but the behaviour was incorrect nonetheless. Instead DetectCondition should be used, if DetectCondition evaluates to false then the package is installed, if true then no action is taken.