Wix4: Fix version strings for not currently installed packages/features was sometimes empty and sometimes 0.0.0.0 depending on how the installed version is determined by the bundle.
When the installed version is determined by searching the registry or searching for a file, if the registry or file is missing then the version is set to an empty string. However if the installed version is determined by looking for an entry in the ARP, then the version is set to 0.0.0.0 if missing. Normalize all of these to null when displaying.