Wix4: Ensure that prerequisite packages are always installed, even if BootstrapperApplication is running.
Wix only checks the prerequisite packages if the bootstrapper application fails to start. but our bootstrapper application might start without all prerequisite packages, e.g. if the x86 version of the desktop runtime is installed but not the x64 version and/or the asp.net core runtime, so make sure these are also installed as part of the remaining installation. Technically these aren't true prerequisites of our bootstrapper but I have chosen to install all .net related packages as prerequisites for simplicity.