MP2-923: Update master_configuration_menu to allow derived screens to define the base layout and menu image directly without relying on the state of the ConfigurationModel
Some screens are based on master_configuration_menu but do not use the ConfigurationModel for navigation, these screens were displayed incorrectly if navigated to directly because the layout of the screen was based on the ConfigurationModel, which could be in an undefined state if the screen was not navigated to via a config section. Instead allow these screens to define the layout and images directly.