MP2-931: Add workaround for namescope issue in ContentControls preventing ElementName bindings from working in EPG style
There is currently an issue with ContentControls, like Button, where if content is added as a child of the ContentControl in xaml and additionally the ContentControl has it's own Template the names of the elements contained in the content are not registered in the correct place. The names get registered in the same namescope as the ContentControl, however the content gets added to the visual tree in the ContentPresenter, which is contained in the namescope of the ContentControl's template, so the element names are not found by any bindings, because they end up looking in the Template namescope and not the ContentControl's namescope.

The workaround is to add the child elements in the ContentControl's ContentTemplate and add the current data context as Content, this causes the element names to be registered in the ContentTemplate namescope rather than the Template or ContentControl's namescope and they are then found correctly