MP2-963: Decrease the TVE3 wait timeout to 15 seconds, partially reverts MP2-946
During installation the setup waits for the MP2 Windows service to start, if it does not start within 30 seconds Windows considers that the service has failed to start and the installation fails with an error. Decrease the TVE3 timeout to less than 30 seconds so that the server does not take longer than 30 seconds to start if TVE3 initializing takes too long.