MP2-944: Set the REINSTALLMODE of the MP2 installation package to amus to force it to always reinstall/overwrite files when installing.
This should fix missing dlls when installing a later version of MP2 with earlier versions of some dlls, e.g. if upgrading a net6 version to a betfx version. The default is to only overwrite older dlls, but this should be safe as MP2 installs to a self-contained directory and doesn't need to worry about interfering with other application installations, which the default REINSTALLMODE helps to prevent.